feat: Refactor Navbar component to enhance functionality and improve structure
