+++ website_nextjs/src/components/layout/navbar/index.ts
@@ -1,6 +1,6 @@
 // Export all components from the navbar folder
-export { default } from './navbar';
-export { default as Navbar } from './navbar';
+export { default } from '../navbar';
+export { Navbar } from '../navbar';
 export { default as Dropdown } from './dropdown';
 export * from './types';
 export * from './config';
